Move struct definitions to include/quoter.hpp
Moving the structure defintions into the header allows
for defining functions that take arguments of the struct
types, as well as cleaning up the source file for the
actual working code. The definitions are still private,
so it's not like anything else will magically see them.

This patch is meant for maintenance purposes - this will
allow future refactors to happen more easily.